feature: added bigdecimal ext

+++ example/src/main/java/com/mercadopago/sdk/android/example/extensions/BigDecimalExt.kt
@@ -1,9 +1,10 @@
 package com.mercadopago.sdk.android.example.extensions
 
+import java.math.BigDecimal
 import java.text.NumberFormat
 import java.util.Locale
 
-fun Float.toCurrencyString(locale: Locale = Locale.getDefault()): String {
+fun BigDecimal.toCurrencyString(locale: Locale = Locale.getDefault()): String {
     val format = NumberFormat.getCurrencyInstance(locale)
     return format.format(this)
 }
